fix: divide foodservicemanager.cs
