+++ Assets/Scripts/Food/DirtyDishService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 90e342d8346e4923ac8dcf36848054d2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
