+++ Assets/Scripts/Food/FoodOrderVisuals.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f4919662b2e4e13aa48104c784aa912
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
