--- Assets/Scripts/Input/ServeInput.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5df9935e870a842c995e2e96a1123c32
